desktop-app/requirements.txt
PyQt5==5.15.9
pycryptodome==3.17
requests==2.28.2
web3==6.1.0
pywin32
